+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [14.2.1] - 2020-05-06
 ### Changed
  - Reverted "PrintVersions" console command back to "VersionInfo"